+++ distributions/univariate/continuous/StudentTRand.h
@@ -11,7 +11,7 @@
  * If X ~ t(1, μ, σ), then X ~ Cauchy(μ, σ)
  * X -> Normal(μ, σ) for t -> ∞
  */
-template < typename RealType = long double >
+template < typename RealType = double >
 class RANDLIBSHARED_EXPORT StudentTRand : public ContinuousDistribution<RealType>
 {
     double nu = 1; ///< degree ν
